+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Fixed
 
+- Improve compatibility with non-IntelliJ IDEs
+
 ## [6.2.1] - 2025-01-05
 
 ### Fixed